Refactor service imports and reorganize service directory structure
