+++ src/components/detail/StorageExpansionDialog.tsx
@@ -21,9 +21,9 @@ import {
 import type { Item } from '../../types/index';
 import useGameStore from '../../store/gameStore';
 import FactorioIcon from '../common/FactorioIcon';
-import { DataService } from '../../services/DataService';
+import { DataService } from '../../services/core/DataService';
 import { getAvailableChestTypes } from '../../data/storageConfigs';
-import { getStorageService } from '../../services/StorageService';
+import { getStorageService } from '../../services/state/StorageService';
 
 interface StorageExpansionDialogProps {
   open: boolean;
