+++ src/components/facilities/PowerManagement.tsx
@@ -19,7 +19,7 @@ import {
   Add,
   Remove
 } from '@mui/icons-material';
-import { PowerService } from '../../services/PowerService';
+import { PowerService } from '../../services/game/PowerService';
 import useGameStore from '../../store/gameStore';
 import FactorioIcon from '../common/FactorioIcon';
