+++ src/components/technology/TechSimpleGrid.tsx
@@ -9,7 +9,7 @@ import {
 import TechGridCard from './TechGridCard';
 import TechVirtualizedGridWithAutoSizer from './TechVirtualizedGridWithAutoSizer';
 import type { Technology, TechStatus } from '../../types/technology';
-import { TechnologyService } from '../../services/TechnologyService';
+import { TechnologyService } from '../../services/game/TechnologyService';
 
 interface TechSimpleGridProps {
   /** 要显示的科技列表 */
