+++ src/data/storageConfigs.ts
@@ -2,7 +2,7 @@
 
 import type { StorageConfig } from '../types/index';
 import { ServiceLocator, SERVICE_NAMES } from '../services/ServiceLocator';
-import type { StorageService } from '../services/StorageService';
+import type { StorageService } from '../services/state/StorageService';
 
 // 重新导出存储配置数据
 export { STORAGE_SPECIFIC_CONFIGS } from './storageConfigData';
