+++ src/hooks/useItemRecipes.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect, useMemo } from 'react';
 import type { Item, Recipe } from '../types/index';
-import { DataService } from '../services/DataService';
-import { RecipeService } from '../services/RecipeService';
+import { DataService } from '../services/core/DataService';
+import { RecipeService } from '../services/core/RecipeService';
 import ManualCraftingValidator from '../utils/manualCraftingValidator';
 
 export const useItemRecipes = (item: Item) => {
