+++ src/hooks/useProductionLoop.ts
@@ -2,9 +2,9 @@
 
 import { useEffect, useCallback, useRef } from 'react';
 import useGameStore from '../store/gameStore';
-import { FuelService } from '../services/FuelService';
-import { RecipeService } from '../services/RecipeService';
-import { PowerService } from '../services/PowerService';
+import { RefactoredFuelService as FuelService } from '../services/game/RefactoredFuelService';
+import { RecipeService } from '../services/core/RecipeService';
+import { PowerService } from '../services/game/PowerService';
 import type { FacilityInstance } from '../types/facilities';
 import { msToSeconds } from '../utils/common';
