+++ src/hooks/useServices.ts
@@ -5,13 +5,13 @@
 
 import { useMemo } from 'react';
 import { ServiceLocator, SERVICE_NAMES } from '../services/ServiceLocator';
-import { DataService } from '../services/DataService';
-import { RecipeService } from '../services/RecipeService';
-import { TechnologyService } from '../services/TechnologyService';
-import { FuelService } from '../services/FuelService';
-import { PowerService } from '../services/PowerService';
-import { StorageService } from '../services/StorageService';
-import { UserProgressService } from '../services/UserProgressService';
+import { DataService } from '../services/core/DataService';
+import { RecipeService } from '../services/core/RecipeService';
+import { TechnologyService } from '../services/game/TechnologyService';
+import { RefactoredFuelService as FuelService } from '../services/game/RefactoredFuelService';
+import { PowerService } from '../services/game/PowerService';
+import { StorageService } from '../services/state/StorageService';
+import { UserProgressService } from '../services/state/UserProgressService';
 import type { GameStateProvider } from '../services/interfaces';
 
 /**
