--- src/services/GameStorageService.ts
@@ -1,2 +1,0 @@
-// Re-export from new location
-export { GameStorageService, gameStorageService } from './state/GameStorageService';
