--- src/services/PowerService.ts
@@ -1,3 +1,0 @@
-// Re-export from new location
-export { PowerService, powerService } from './game/PowerService';
-export type { PowerBalance, FacilityPowerInfo } from './game/PowerService';
