--- src/services/RecipeService.ts
@@ -1,2 +1,0 @@
-// Re-export from new location
-export { RecipeService } from './core/RecipeService';
