--- src/services/ServiceInitializer.ts
@@ -1,2 +1,0 @@
-// Re-export from new location
-export { ServiceInitializer } from './utils/ServiceInitializer';
