--- src/services/StorageService.ts
@@ -1,3 +1,0 @@
-// Re-export from new location
-export { StorageService, storageService, getStorageService } from './state/StorageService';
-export type { StorageConfig } from '../types/index';
