--- src/services/TechnologyService.ts
@@ -1,3 +1,0 @@
-// Re-export from new location
-export { TechnologyService, technologyService } from './game/TechnologyService';
-export type { Technology, TechStatus, TechTreeState, ResearchQueueItem, TechResearchState } from '../types/technology';
